Don't use Python's built-in PRNG to generate session ID
Jeremy Allen discovered and reported a place where we use Python's built-in
PRNG to generate session IDs. Neatx' current design doesn't give the session ID
a lot of trust and hence this is a low-risk issue.

The built-in PRNG is seeded from /dev/urandom, which can be considered a good
seed in this case.

Session shadowing between different users, for which we'd have to give session
IDs more trust, is not implemented, and the session directories are all owned
by the logged in user and have the permissions 0700. Session directories are
created using a safe algorithm. Unless one knows the password of another user,
in which case everything is lost, a session ID alone is of no use with Neatx.

Signed-off-by: Michael Hanselmann <hansmi@google.com>
Reviewed-by: René Nussbaumer <rn@google.com>